feat(core/system): rename legacy modules/ to system/
